chore (api): restructure to use request util
